Moved effect classes to their own namespace.
